Lists and hooks
lifecycle class component : componentDidMount  is called after the component has been rendered to the DOM.shouldComponentUpdate method is called before the component re-renders.componentDidUpdate method is called after the component has been updated.

lifecycle functional  component :

useEffect with an empty dependency array is equivalent to componentDidMount, and the second useEffect with a dependency array containing count is equivalent to componentDidUpdate. The cleanup function in the first useEffect is equivalent to componentWillUnmount.